Fix LoopChoiceNode::EatsAtLeastFromLoopEntry()
In https://crrev.com/c/5332375 we changed RegexpNode::EatsAtLeast to
return a uint32_t (as it represents a string length). However,
EatsAtLeastFromLoopEntry() relied on signed integer underflows (followed
by a saturation cast that converts negative numbers to zero). This CL
now fixes that logic again by using an explicit cast to int.

Drive-By: Also fix sampling logic in RegExpImpl::Compile.

Bug: 374400655
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5946967
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96746}

+++ src/regexp/regexp-compiler.cc
@@ -1534,10 +1534,13 @@ EatsAtLeastInfo LoopChoiceNode::EatsAtLeastFromLoopEntry() {
   // successful match in the loop body must also include the continuation node.
   // However, in some cases involving positive lookaround, the loop body under-
   // reports its appetite, so use saturated math here to avoid negative numbers.
+  // For this to work correctly, we explicitly need to use signed integers here.
   uint8_t loop_body_from_not_start = base::saturated_cast<uint8_t>(
-      loop_node_->EatsAtLeast(true) - continue_node_->EatsAtLeast(true));
+      static_cast<int>(loop_node_->EatsAtLeast(true)) -
+      static_cast<int>(continue_node_->EatsAtLeast(true)));
   uint8_t loop_body_from_possibly_start = base::saturated_cast<uint8_t>(
-      loop_node_->EatsAtLeast(false) - continue_node_->EatsAtLeast(true));
+      static_cast<int>(loop_node_->EatsAtLeast(false)) -
+      static_cast<int>(continue_node_->EatsAtLeast(true)));
 
   // Limit the number of loop iterations to avoid overflow in subsequent steps.
   int loop_iterations = base::saturated_cast<uint8_t>(min_loop_iterations());
